+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/BreadcrumbPriorityQueue.kt
@@ -4,6 +4,7 @@
 
 package mozilla.components.lib.crash
 
+import mozilla.components.support.base.crash.Breadcrumb
 import java.util.PriorityQueue
 
 internal class BreadcrumbPriorityQueue(