+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/service/CrashReporterService.kt
@@ -4,8 +4,8 @@
 
 package mozilla.components.lib.crash.service
 
-import mozilla.components.lib.crash.Breadcrumb
 import mozilla.components.lib.crash.Crash
+import mozilla.components.support.base.crash.Breadcrumb
 
 internal const val INFO_PREFIX = "[INFO]"